--- pkg/handlers/handler-functions.go
@@ -1,87 +1,0 @@
-package handlers
-
-import (
-	"log"
-	"net/http"
-)
-
-func HandleRoot(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.Write([]byte("{}"))
-}
-
-func HandleUsers(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	log.Println(r.URL.Path)
-	handlers.HandleListUsers(w, r)
-}
-
-func HandleRegistration(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleRegUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleLogin(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleLoginUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleLogout(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleLogoutUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleProfileData(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleEditProfileUserData(w, r)
-		return
-	}
-	if r.Method == http.MethodGet {
-		handlers.HandleGetProfileUserData(w, r)
-		return
-	}
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleProfilePicture(w http.ResponseWriter, r *http.Request) {
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleEditProfileUserPicture(w, r)
-		return
-	}
-	if r.Method == http.MethodGet {
-		handlers.HandleGetProfileUserPicture(w, r)
-		return
-	}
-	handlers.HandleEmpty(w, r)
-}
